requirements.txt
elasticsearch==8.15.1
python-dotenv==1.0.1
pandas==2.2.3
rich==13.9.2
nltk==3.8.1
transformers==4.45.2
